% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{cleanMem}
\alias{cleanMem}
\title{Clean memory.}
\usage{
cleanMem(n = 10)
}
\description{
Check \url{https://stackoverflow.com/questions/1467201/forcing-garbage-collection-to-run-in-r-with-the-gc-command}
}
\examples{
\dontrun{
cleanMem()
}
}
