% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{intensity_per_frame}
\alias{intensity_per_frame}
\title{Get sum of intensity per each frame (retention time).}
\usage{
intensity_per_frame(timsr, recalibrated = TRUE)
}
\arguments{
\item{timsr}{Instance of TimsR}

\item{recalibrated}{Use Bruker recalibrated total intensities or calculate them from scratch with OpenTIMS?}
}
\description{
Get sum of intensity per each frame (retention time).
}
