% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{retention_times}
\alias{retention_times}
\title{Get the retention time for each frame.}
\usage{
retention_times(timsr)
}
\arguments{
\item{timsr}{Instance of TimsR.}
}
\value{
Retention times corresponding to each frame.
}
\description{
Get the retention time for each frame.
}
