% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rol_hor_functions.R
\name{rolling_horizon}
\alias{rolling_horizon}
\title{Rolling horizon}
\usage{
rolling_horizon(series, FUN, FUN_PRED, start_eval, eval_periods = 4,
  h = 14, regressors = NULL)
}
\description{
Test time series models with the rolling horizon method.
}
\details{
@param series Time series of the target variable.
 @param FUN Training function, returns a model-object.
 @param FUN_PRED Prediction function, returns a prediction.
 @param start_eval Vector in the shape c(year, period) that form the last
    point of the first rolling horizon.
 @param eval_periods Number of rolling horizons.
 @param h Horizon length.
}
