% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{[,TimsR,ANY,ANY-method}
\alias{[,TimsR,ANY,ANY-method}
\title{Get some frames of data.}
\usage{
\S4method{[}{TimsR,ANY,ANY}(x, i)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{i}{An array of nonzero indices to extract.}
}
\description{
Get some frames of data.
}
\examples{
\dontrun{
D = TimsR('path/to/your/folder.d')
print(head(D[10]))
print(head(D[10:100]))
}
}
