% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{table2dt}
\alias{table2dt}
\title{Extract tables from sqlite database analysis.tdf.}
\usage{
table2dt(timsr, names)
}
\arguments{
\item{timsr}{Instance of TimsR}

\item{names}{Names to extract from the sqlite database.}
}
\value{
A data.table with proper contents.
}
\description{
Export a table from sqlite.
}
