% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{tables_names}
\alias{tables_names}
\title{Extract tables from sqlite database analysis.tdf.}
\usage{
tables_names(timsr)
}
\arguments{
\item{timsr}{Instance of TimsR}
}
\value{
Names of tables.
}
\description{
Extract tables from sqlite database analysis.tdf.
}
