% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{tdf.tables}
\alias{tdf.tables}
\title{Explore the contentents of the sqlite .tdf database.}
\usage{
tdf.tables(timsr)
}
\arguments{
\item{timsr}{Instance of TimsR}
}
\value{
List of data.tables filled with data from 'analysis.tdf'.
}
\description{
Explore the contentents of the sqlite .tdf database.
}
\examples{
\dontrun{
D = TimsR('path/to/your/folder.d')
print(tdf.tables(D)) 
}
}
