% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{TimsR}
\alias{TimsR}
\title{Get TimsR data accessor.}
\usage{
TimsR(path.d)
}
\arguments{
\item{path.d}{Path to the TimsTOF '*.d' folder containing the data (requires the folder to contain only 'analysis.tdf' and 'analysis.tdf_bin').}
}
\description{
Get TimsR data accessor.
}
\examples{
\dontrun{
D = TimsR(path_to_.d_folder)
D[1] # First frame.
}
}
