% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{plot_TIC}
\alias{plot_TIC}
\title{Plot intensity per retention time.}
\usage{
plot_TIC(timsr, recalibrated = TRUE)
}
\arguments{
\item{timsr}{Instance of TimsR}

\item{recalibrated}{Use Bruker recalibrated total intensities or calculate them from scratch with OpenTIMS?}
}
\description{
Plot will split 'MS1' and 'MS2'.
}
