% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsr.R
\name{[,TimsR,ANY,character-method}
\alias{[,TimsR,ANY,character-method}
\title{Get some frames of data.}
\usage{
\S4method{[}{TimsR,ANY,character}(x, i, j)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{i}{An array of nonzero indices to extract.}
}
\description{
Get some frames of data.
}
\examples{
\dontrun{
D = TimsR('path/to/your/folder.d')
print(D[c(1,20, 53), c('frame','scan','tof','intensity','mz','inv_ion_mobility','retention_time')] # extract all columns
print(D[c(1,20, 53), c('scan','intensity')] # only 'scan' and 'intensity'
}
}
